Rename compose.rs to composition.rs for consistency
- Renamed module file from compose.rs to composition.rs
- Updated module declarations in main.rs
- Updated imports in identity.rs tests
